任务：小组功能实现

+++ server/src/main/java/zheye/server/entity/Support.java
@@ -6,7 +6,8 @@ import lombok.Data;
 public class Support {
     private int userId;
     private int supportCommentId;
-    private int id;
+
+    private Long id;
 
     @Override
     public String toString() {
